WIP command bar and list view for vehicles
I'm not sure exacly how I want it to look. To start with, I will
have a command bar at the top and a list view for the vehicles
below it. For now, there is a hack to put an item in the list so
that something is displayed when the app is run.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>