Add vehicle info view
Created the vehicle info view. Laid out simple controls for the view,
including all data input fields, a save button, and a cancel button.
Save will create a new Vehicle object and add it to the Vehicles table
in the database.

This is a simple implementation of the new vehicle data entry. The form
does not look great. User can still not edit or delete an existing
vehicle.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>