Implement delete vehicle
The Delete Vehicle button now deletes the selected vehicle.

This required changing the database model to make the relationship from the
Purchase to the Vehicle required instead of optional. This was done by adding
the PK of the Vehicle to the Purchase. This resulted in the migration having
to do a Remove Primary Key operation (or something similar), which SQLite
doesn't support. The app would throw this exception every time it was executed,
even if the app didn't exist (and thus there should have been no database to
modify). My solution was to delete all of the migrations, then do an Add-
Migration. Since it was as if this was the first time it had been done, there
was no need for the PK modification, and the exception was avoided.

The delete command does not ask the user for confirmation, which would
probably be a nice addition for usability.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>