Simple sample data in purchase list
I added two new purchase list views along the bottom.

In the one on the left, I am doing the "simple" sample data method
of using x:Bind to a collection object. The designer populates this
with three items to give an idea of what it will look like.

To support this, I created two new classes, PurchasesViewModel
and PurchasesCollection. The collection simply inherits from
ObservableCollection<Purchase>. The view model has an object of
this type and a get property for it. For now, a single item is
added to the collection manually in the constructor. Later I will
try to get it poplulated from the database.

I have not done anything yet with the list on the right. It
will be used to show the Binding method that actually uses sample
data from a file.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>