WIP on the second test list
I got it set up with a sample data file, but then I realized
I had used the wrong collection. I don't want to use Purchase,
I want to use PurchaseSummary, since that has the Mpg and Dpg.
Also, I realized that not all of the properties were given values.
Specifically, Cost and Gallons were both 0, which are the only
things I display, so it made it look like it was the same item
over and over again.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>