Second purchases list works with new view model
I fixed the sample data issue by generating a file based off of the correct
class. However, the generation only auto-populates basic types like string
and DateTime. This means that the decimal values are blank. I spent more
than an hour searching to find out how to specify these values in the XAML
file but I couldn't find anything. Thus, my conclusion is that I need to make
everything a string in the "view model" class. This seems like an
unnecessary extra step, but I guess it needs to be done at some point. I
will make this change in a future commit.

The ViewModel member is now modified when the user selects a vehicle, instead
of setting the ListView's ItemsSource directly. I am not exactly sure how it
works, but the change is picked up by the GUI and the list works as expected.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>