Convert remaining members of PurchasesSummary to string
Made everything in PurchasesSummary a string, then generated a new
sample data file so that everything would be seen in the designer.
I manually modified the first element in the file to be more
realistic, using what I thought would be the "widest" numbers a user
would be likely to encounter. For some reason the designer is not making
the column wide enought to see everything. Maybe it cannot do this even
though it has the sample data; the width seems to be based on a TextBlock
with a static string.

I think I am done with this branch. I am not going to merge it into master
because I think there is some junk in there, but I am not going to delete it
or squash it because I want the history to be there for reference.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>