Refactor PurchaseSummary to use all strings
From my research on binding sample data, I have decided that it is
easier to have the view object be all strings instead of the complex
types of the model. This requires a conversion in the code, but I
don't mind writing it. This method might also violate the spirit of
the view model, but I am willing to live with it until I figure out
a better way.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>