Add sample data for purchases view
I used Blend to generate the sample data file, then manually edited
it to give it reasonable values. A one-line change to the binding
resource is all that was needed to get the designer to use the sample
data in the view.

I haven't decided whether or not to do this for the vehicles. That list
is much simpler so I might postpone that until other more important things
are finished.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>